initial apis for the admin
