+++ .gitignore
@@ -1,3 +1,4 @@
 env
 .vscode
 __pycache__
+admin/web/__pycache__
