+++ admin/web/admin.py
@@ -1,55 +1,31 @@
-from fastapi.routing import APIRouter 
+from fastapi.routing import APIRouter
+from .strategy import router as strategy_router
+from .users_management import router as users_router
+from .keys import router as keys_router
 
 router = APIRouter()
+router.include_router(users_router, prefix="/users")
+router.include_router(strategy_router, prefix="/strategy")
+router.include_router(keys_router, prefix="/key")
+
 
 @router.get("/login")
 async def login():
-    '''login and redirect the user to the dashboard'''
+    """login and redirect the user to the dashboard"""
     pass
+
 
 @router.get("/logout")
 async def logout():
-    '''logout and redirect the user to the login route'''
+    """logout and redirect the user to the login route"""
     pass
+
 
 # ! Todo: is this realy necessary in case of using react ?
 @router.get("/dashboard")
 async def dashboard():
-    '''show the dashboard'''
-    pass
-
-# ----- Users Management -------
-@router.post("/addUser")
-async def add_user():
-    '''create a new user and store him in the DB'''
-    pass
-
-@router.delete("/deleteUsers")
-async def remove_users(users_id: list[int]):
-    '''remove user completely from the DB'''
-    pass
-
-@router.put("/updateUser")
-async def update_user():
-    '''update user info'''
+    """show the dashboard"""
     pass
 
 
-@router.get("/user/{id}")
-async def get_user(id):
-    '''return user using his id'''
-    pass
-
-@router.get("/users")
-async def get_all_users():
-    '''return all users from the DB.
-    results will be sent to the front-end to filer them'''
-    pass
-
-
-async def create_api_key():
-    '''generate an api key for a user'''
-    pass
-
-# ----- Strategy Management -------
